+++ indexer/kv_indexer.go
@@ -57,7 +57,7 @@ func (kv *KVIndexer) IndexBlock(block *tmtypes.Block, txResults []*abci.Response
 	var ethTxIndex int32
 	for txIndex, tx := range block.Txs {
 		result := txResults[txIndex]
-		if !rpctypes.TxSuccessOrExceedsBlockGasLimit(result) {
+		if !rpctypes.TxSucessOrExpectedFailure(result) {
 			continue
 		}